fix(cuda): cast root-sum doubles at the HistCell boundary
The one f32 narrowing site only a BONSAI_CUDA=ON build reaches; caught by
the L40S validation build, missed by CI's cuda-compile job.

+++ src/cuda/histogram_engine.cu
@@ -571,7 +571,8 @@ bool CudaHistogramEngine::begin_root(Dataset const &ds, floats_view grad,
         sg += grad[r];
         sh += hess[r];
     }
-    root.sums      = {.sum_grad = sg, .sum_hess = sh};
+    root.sums      = {.sum_grad = static_cast<float>(sg),
+                      .sum_hess = static_cast<float>(sh)};
     root.row_count = root.rows.size();
     if (im.prof_counters.enabled)
     {
